requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
pydantic==2.13.1
numpy==2.4.4
pandas==3.0.2
scikit-learn==1.8.0
sentence-transformers==5.4.1
faiss-cpu==1.13.2
pytest==9.0.3
httpx==0.28.1
joblib==1.5.3
xgboost==3.2.0
lightgbm==4.6.0